content/base/test/test_bug498433.html

<!DOCTYPE HTML>
<html>
<!--
-->
<head>
  <title>Test for HTML serializer</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=498433">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe id="testframe" src="file_htmlserializer_ipv6.html">
  </iframe>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">


function loadFileContent(aFile, aCharset) {
    if (aCharset == undefined)
        aCharset = 'UTF-8';

    var baseUri = Components.classes['@mozilla.org/network/standard-url;1']
                   .createInstance(Components.interfaces.nsIURI);
    baseUri.spec = window.location.href;

    var ios = Components.classes['@mozilla.org/network/io-service;1']
            .getService(Components.interfaces.nsIIOService);
    var chann = ios.newChannel(aFile, aCharset, baseUri);

    var cis = Components.interfaces.nsIConverterInputStream;

    var inputStream = Components.classes["@mozilla.org/intl/converter-input-stream;1"]
                       .createInstance(cis);
    inputStream.init(chann.open(), aCharset, 1024, cis.DEFAULT_REPLACEMENT_CHARACTER);
    var str = {}, content = '';
    while (inputStream.readString(4096, str) != 0) {
        content += str.value;
    }
    return content;
}

function isRoughly(actual, expected, message) {
  return is(actual.replace("<!DOCTYPE HTML", "<!DOCTYPE html"), 
            expected, 
            message);
}

function testHtmlSerializer_1 () {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  const de = Components.interfaces.nsIDocumentEncoder
  var encoder = Components.classes["@mozilla.org/layout/documentEncoder;1?type=text/html"]
                   .createInstance(Components.interfaces.nsIDocumentEncoder);

  var doc = $("testframe").contentDocument;
  var out, expected;

  // in the following tests, we must use the OutputLFLineBreak flag, to avoid
  // to have the default line break of the platform in the result, so the test
  // can pass on all platform

  //------------ no flags
  encoder.init(doc, "text/html", de.OutputLFLineBreak);
  encoder.setCharset("UTF-8");
  out = encoder.encodeToString();
  expected = loadFileContent("file_htmlserializer_ipv6_out.html");
  isRoughly(out, expected, "test no flags");
  //------------ OutputAbsoluteLinks
  encoder.init(doc, "text/html", de.OutputLFLineBreak | de.OutputAbsoluteLinks);
  encoder.setCharset("UTF-8");
  out = encoder.encodeToString();
  expected = loadFileContent("file_htmlserializer_ipv6_out.html");
  isRoughly(out, expected, "test OutputAbsoluteLinks");
  //------------ serializing a selection
  encoder.init(doc, "text/html", de.OutputLFLineBreak | de.OutputAbsoluteLinks);
  encoder.setNode(doc.links[0]);
  out = encoder.encodeToString();
  expected = "<a href=\"http://[2001:4860:a003::68]/\">Test</a>";
  isRoughly(out, expected, "test selection");


  SimpleTest.finish();
}


SimpleTest.waitForExplicitFinish();

addLoadEvent(testHtmlSerializer_1);

</script>
</pre>

</body>
</html>
